Replace the benchmark with the regime matrix
Drop-in: same invocation, same report heading the sticky-comment step matches
on, same dnsmasq port the workflow starts. The old single-regime script is gone
rather than left alongside, since its one number is the thing that could not
detect a regression.

Clients are now warmed before timing. Socket setup and probing cost the same
whatever the query count, so leaving it inside the measurement made two runs at
different sizes incomparable: the persistent pool regime read 3,694 qps at 4,000
queries and 16,485 at 20,000 with no config change.
